Trait actix_jwt_authc::InvalidatedJWTsReader

pub trait InvalidatedJWTsReader<T = String> {
    fn read<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        tag: Option<&'life1 InvalidatedTokensTag<T>>
    ) -> Pin<Box<dyn Future<Output = Result<InvalidatedTokens<T>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

A reader for invalidated JWTs, for instance from an external shared data store.

Required Methods

fn read<'life0, 'life1, 'async_trait>(
&'life0 self,
tag: Option<&'life1 InvalidatedTokensTag<T>>
) -> Pin<Box<dyn Future<Output = Result<InvalidatedTokens<T>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Reads invalidated JWTs

The read method will receive the last InvalidatedTokensTag that the middleware received as an argument, and must return either a variant of InvalidatedTokens or an IO error.